move display and summary to trainer
