+++ pykg2vec/core/Rescal.py
@@ -156,80 +156,3 @@ class Rescal(ModelMeta):
     def get_proj_embed(self, h, r, t, sess=None):
         """function to get the projectd embedding value in numpy"""
         pass
-
-    def summary(self):
-        """function to print the summary"""
-        print("\n----------------SUMMARY----------------")
-        # Acquire the max length and add four more spaces
-        maxspace = len(max([k for k in self.config.__dict__.keys()])) + 15
-        for key, val in self.config.__dict__.items():
-            if 'gpu' in key:
-                continue
-            if len(key) < maxspace:
-                for i in range(maxspace - len(key)):
-                    key = ' ' + key
-            print(key, ":", val)
-        print("-----------------------------------------")
-
-    def display(self, sess=None):
-        """function to display embedding"""
-        if self.config.plot_embedding:
-            triples = self.data_handler.validation_triples_ids[:self.config.disp_triple_num]
-            viz = Visualization(triples=triples,
-                                idx2entity=self.data_handler.idx2entity,
-                                idx2relation=self.data_handler.idx2relation)
-
-            viz.get_idx_n_emb(model=self, sess=sess)
-            viz.reduce_dim()
-            viz.plot_embedding(resultpath=self.config.figures, algos=self.model_name)
-
-        if self.config.plot_training_result:
-            viz = Visualization()
-            viz.plot_train_result(path=self.config.result,
-                                  result=self.config.figures,
-                                  algo=['TransE', 'TransR', 'TransH'],
-                                  data=['Freebase15k'])
-
-        if self.config.plot_testing_result:
-            viz = Visualization()
-            viz.plot_test_result(path=self.config.result,
-                                 result=self.config.figures,
-                                 algo=['TransE', 'TransR', 'TransH'],
-                                 data=['Freebase15k'], paramlist=None, hits=self.config.hits)
-
-def main(_):
-    parser = ArgumentParser(description='Knowledge Graph Embedding with Rescal')
-    parser.add_argument('-b', '--batch', default=128, type=int, help='batch size')
-    parser.add_argument('-t', '--tmp', default='../intermediate', type=str, help='Temporary folder')
-    parser.add_argument('-ds', '--dataset', default='Freebase15k', type=str, help='Dataset')
-    parser.add_argument('-l', '--epochs', default=1, type=int, help='Number of Epochs')
-    parser.add_argument('-tn', '--test_num', default=100, type=int, help='Number of test triples')
-    parser.add_argument('-ts', '--test_step', default=5, type=int, help='Test every _ epochs')
-    parser.add_argument('-lr', '--learn_rate', default=0.1, type=float, help='learning rate')
-    parser.add_argument('-gp', '--gpu_frac', default=0.4, type=float, help='GPU fraction to use')
-    parser.add_argument('-k', '--embed', default=50, type=int, help='Hidden embedding size')
-    args = parser.parse_args()
-
-    if not os.path.exists(args.tmp):
-        os.mkdir(args.tmp)
-
-    data_handler = DataPrep(args.dataset)
-    args.test_num = min(len(data_handler.test_triples_ids), args.test_num)
-    
-    config = RescalConfig(learning_rate=args.learn_rate,
-                          batch_size=args.batch,
-                          epochs=args.epochs,
-                          hidden_size=args.embed)
-
-    config.test_step = args.test_step
-    config.test_num  = args.test_num
-    config.gpu_fraction = args.gpu_frac
-
-    model = Rescal(config, data_handler)
-    
-    trainer = Trainer(model=model)
-    trainer.build_model()
-    trainer.train_model()
-
-if __name__ == "__main__":
-    tf.app.run()
