+++ pykg2vec/core/SMELinear.py
@@ -136,43 +136,3 @@ class SMELinear(ModelMeta):
     def get_proj_embed(self, h, r, t, sess=None):
         """function to get the projectd embedding value in numpy"""
         pass
-
-    def summary(self):
-        """function to print the summary"""
-        print("\n----------------SUMMARY----------------")
-        # Acquire the max length and add four more spaces
-        maxspace = len(max([k for k in self.config.__dict__.keys()])) + 15
-        for key, val in self.config.__dict__.items():
-            if 'gpu' in key:
-                continue
-            if len(key) < maxspace:
-                for i in range(maxspace - len(key)):
-                    key = ' ' + key
-            print(key, ":", val)
-        print("-----------------------------------------")
-
-    def display(self, sess=None):
-        """function to display embedding"""
-        if self.config.plot_embedding:
-            triples = self.data_handler.validation_triples_ids[:self.config.disp_triple_num]
-            viz = Visualization(triples=triples,
-                                idx2entity=self.data_handler.idx2entity,
-                                idx2relation=self.data_handler.idx2relation)
-
-            viz.get_idx_n_emb(model=self, sess=sess)
-            viz.reduce_dim()
-            viz.plot_embedding(resultpath=self.config.figures, algos=self.model_name)
-
-        if self.config.plot_training_result:
-            viz = Visualization()
-            viz.plot_train_result(path=self.config.result,
-                                  result=self.config.figures,
-                                  algo=['TransE', 'TransR', 'TransH'],
-                                  data=['Freebase15k'])
-
-        if self.config.plot_testing_result:
-            viz = Visualization()
-            viz.plot_test_result(path=self.config.result,
-                                 result=self.config.figures,
-                                 algo=['TransE', 'TransR', 'TransH'],
-                                 data=['Freebase15k'], paramlist=None, hits=self.config.hits)
