+++ pykg2vec/core/TransE.py
@@ -127,42 +127,3 @@ class TransE(ModelMeta):
     def get_proj_embed(self, h, r, t, sess=None):
         """function to get the projected embedding value in numpy"""
         return self.get_embed(h, r, t, sess)
-
-    def summary(self):
-        """function to print the summary"""
-        print("\n----------------SUMMARY----------------")
-        # Acquire the max length and add four more spaces
-        maxspace = len(max([k for k in self.config.__dict__.keys()])) + 15
-        for key, val in self.config.__dict__.items():
-            if 'gpu' in key:
-                continue
-            if len(key) < maxspace:
-                for i in range(maxspace - len(key)):
-                    key = ' ' + key
-            print(key, ":", val)
-        print("-----------------------------------------")
-
-    def display(self, sess=None):
-        """function to display embedding"""
-        if self.config.plot_embedding:
-            viz = Visualization(model=self,
-                                ent_only_plot=True,
-                                rel_only_plot=True,
-                                ent_and_rel_plot=True)
-
-            viz.plot_embedding(sess=sess, resultpath=self.config.figures, algos=self.model_name,
-                               show_label=False)
-
-        if self.config.plot_training_result:
-            viz = Visualization()
-            viz.plot_train_result(path=self.config.result,
-                                  result=self.config.figures,
-                                  algo=['TransE', 'TransR', 'TransH'],
-                                  data=['Freebase15k'])
-
-        if self.config.plot_testing_result:
-            viz = Visualization()
-            viz.plot_test_result(path=self.config.result,
-                                 result=self.config.figures,
-                                 algo=['TransE', 'TransR', 'TransH'],
-                                 data=['Freebase15k'], paramlist=None, hits=self.config.hits)
